+++ app/Http/Requests/DirtRally/ChampionshipSeasonEventStageRequest.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\DirtRally;
 
-class SeasonEventStageRequest extends Request
+use App\Http\Requests\Request;
+
+class ChampionshipSeasonEventStageRequest extends Request
 {
     /**
      * Get the validation rules that apply to the request.
