+++ app/Http/Requests/DirtRally/ChampionshipSeasonRequest.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\DirtRally;
 
-class ChampionshipRequest extends Request
+use App\Http\Requests\Request;
+
+class ChampionshipSeasonRequest extends Request
 {
     /**
      * Get the validation rules that apply to the request.
